008 impl blocks and associated functions
We're going to spawn more tiles in with the same logic we used to place the placeholders so we'll refactor that logic out into the `Board` 's associated functions using an `impl` block.

We'll create two functions: `Board::new` and `Board::grid_to_world_position`, as well as creating a few new fields including `world_size` on `Board`. The fields `tile_size` and `tile_spacer` are the values we previous created in the `spawn_board` function.

```rust
\#[derive(Resource, Clone)]
struct Board {
    size: u16,
    world_size: f32,
    tile_size: f32,
    tile_spacer: f32,
}

impl Board {
    fn new(size: u16) -> Self {
        let tile_size: f32 = 40.0;
        let tile_spacer: f32 = 10.0;

        let world_size = f32::from(size) * tile_size
            + f32::from(size + 1) * tile_spacer;
        Board {
            size,
            world_size,
            tile_size,
            tile_spacer,
        }
    }
    fn grid_to_world_position(&self, pos: u16) -> f32 {
        let offset = -self.world_size / 2.0
            + 0.5 * self.tile_size;

        offset
            + f32::from(pos) * self.tile_size
            + f32::from(pos + 1) * self.tile_spacer
    }
}
```

`Board::new` will be responsible for creating a new `Board` struct for us. The pattern of defining a `new` method is fairly common in the Rust ecosystem. We can take advantage of this to add an additional `world_size` field on the struct, even though the user will never need to know how it's created.

`Self` in this case refers to `Board` because that's the type we're implementing methods for.

`grid_to_world_position` is a copy/paste of the logic we've already defined in earlier lessons. The biggest difference is that we're now using `self.world_size` because we are going to call this method on an instantiated struct, which means we have access to the fields on the `Board` struct. `grid_to_world_position` will turn a grid position like `0`,`1`,`2`, or `3` into a physical position like `10.0`, `50.0`, etc.

In our `Resource` we can now instantiate the board with `new()` instead of a struct.

```rust
.insert_resource(Board::new(4))
```

and we can remove the old `board_world_size`, `tile_size`, and `tile_spacer` in `spawn_board`, replacing it with using `board.world_size`, `board.tile_size`, and `board.tile_spacer` when spawning our board.

```rust
custom_size: Some(Vec2::splat(
    board.world_size,
)),
```

Our `Transform` component for the tile_placeholders can now use the board to translate an x or y cell position into a world position, cleaning up our code quite a bit by removing the math bits floating around, and giving a name to what we're trying to do in each place.

```rust
transform: Transform::from_xyz(
    board.grid_to_world_position(tile.0),
    board.grid_to_world_position(tile.1),
    1.0,
),
```

remember to remove the `offset` from `.with_children` as well, since we aren't using it anymore.

![full board](https://res.cloudinary.com/dilgcuzda/image/upload/v1727755688/workshops/2048-with-bevy-ecs/bevy-0.14/007-tile-placeholders-with-spacing.avif)