009 - Spawning tiles with rand's choose_multiple
To start a 2048 game we need to spawn 2 tiles onto the board. To do that we're going to create a new startup system that will handle picking board positions and spawning the tiles.

\## New Components

We're going to create two new `Component`s to store value and position data for our entities.

Our first component, `Points` will hold the value for a tile. In 2048, that value is in the sequence 2^n: 2,4,8,16,32,etc because our tiles start at 2 and combine with tiles of the same point value.

```rust
\#[derive(Component)]
struct Points {
    value: u32,
}
```

Our second component holds the position of the tile on the grid with an `x` and a `y` value. Positions of cells in our grid are `u16`s and Bevy's math module already contains a struct where `x` and `y` are `u16`s. Its called `bevy::math::U16Vec2`. We want to be able to query for this component later, so we'll wrap it in our own `Position` newtype. This means we can query Bevy for `Position`s while keeping our data in a `U16Vec2`.

```rust
\#[derive(Component)]
struct Position(U16Vec2);
```

\## The spawn_tiles System

The `spawn_tiles` system is going to be responsible for spawning two tiles in randomized positions to start the game.

Be sure to add the `rand` crate using cargo add so we can use it for randomness

```rust
cargo add rand@0.8.5
```

and bring the `rand` prelude in at the top of our file.

```
use bevy::prelude::*;
use itertools::Itertools;
use rand::prelude::*;
```

The system will need two arguments:

- `Commands` to spawn tile sprites
- `Res<Board>` to get the board size

```rust
fn spawn_tiles(mut commands: Commands, board: Res<Board>) {
    let mut rng = rand::thread_rng();
    let starting_tiles: Vec<(u16, u16)> = (0..board.size)
        .cartesian_product(0..board.size)
        .choose_multiple(&mut rng, 2);

    for (x, y) in starting_tiles.into_iter() {
        let pos = Position(U16Vec2::new(x, y));
        commands.spawn((
            SpriteBundle {
                sprite: Sprite {
                    color: Color::srgb(0.54, 0.64, 0.72),
                    custom_size: Some(Vec2::splat(
                        board.tile_size,
                    )),
                    ..default()
                },
                transform: Transform::from_xyz(
                    board.grid_to_world_position(pos.0.x),
                    board.grid_to_world_position(pos.0.y),
                    1.0,
                ),
                ..default()
            },
            Points { value: 2 },
            pos,
        ));
    }
}
```

After generating all of the possible grid locations, the random number generator will allow us to choose two of them, semi-randomly.

`starting_tiles` will be our two tiles to spawn on the board. We store them in a `Vec` as a two-tuple of `u16` numbers. We already talked about using `.cartesian_product` to generate all the grid tiles and we do the same here.

After generating the iterator of `(u16, u16)`, we can chain `[.choose_multiple](https://docs.rs/rand/0.8.5/rand/seq/trait.IteratorRandom.html#method.choose_multiple)` to use the random number generator to pick two of them.

```rust
let mut rng = rand::thread_rng();
let starting_tiles: Vec<(u16, u16)> = (0..board.size)
    .cartesian_product(0..board.size)
    .choose_multiple(&mut rng, 2);
```

with the `starting_tiles` picked out, we can iterate over the tuples and insert each of the titles. First we'll destructure the `x` and `y` values for each item using a for loop.

After destructuring the values we can create a new `Position` component with the `x` and `y` values. We do this early in the functions because it makes it a bit more clear when we use it later on.

We can iterate over `starting_tiles` with `into_iter` which will give us access to the `u16` values.

```rust
for (x, y) in starting_tiles.iter() {
    let pos = Position(U16Vec2::new(x, y));
```

Spawning a `SpriteBundle` using `commands` is the same as spawning the board and the tile placeholders. The sprite will be `board.tile_size` width and height size. The color will be new, and we fill in the rest of the fields with the `default()`s.

```rust
commands.spawn((
    SpriteBundle {
        sprite: Sprite {
            color: Color::srgb(0.84, 0.89, 0.93),
            custom_size: Some(Vec2::splat(
                board.tile_size,
            )),
            ..default()
        },
        transform: Transform::from_xyz(
            board.grid_to_world_position(pos.0.x),
            board.grid_to_world_position(pos.0.y),
            1.0,
        ),
        ..default()
    },
    Points { value: 2 },
    pos,
));
```

Which leaves us with the `Transform`. The `Transform` will be the same calculation that we used to place the placeholders, so we can use `Board::grid_to_world_position` to find the right place for them.

Finally we insert the `Points` component with a default value of 2 and the `Position` component we created earlier.

and then we need to run the `spawn_tiles` system at startup, alongside `setup` and `spawn_board`.

```rust
.add_systems(Startup, (
    setup,
	spawn_board,
	spawn_tiles
))
```

and every time we re-run our app, we get new tiles on the board.

![first run](https://res.cloudinary.com/dilgcuzda/image/upload/v1727760516/workshops/2048-with-bevy-ecs/bevy-0.14/009-starting-tiles-1.avif)

![second run](https://res.cloudinary.com/dilgcuzda/image/upload/v1727760533/workshops/2048-with-bevy-ecs/bevy-0.14/009-starting-tiles-2.avif)