10 - Displaying Text on a Tile
We want to show the values from the `Points` component on each tile, so we are going to start with showing any text at all on each tile.

We'll create a unit struct, that is: a struct with no fields, to tag the text so we can find it later.

```rust
\#[derive(Component)]
struct TileText;
```

We'll put `.with_children` right after the code we already wrote to spawn new tiles.

The tile sprite children will be a [`Text2dBundle`](https://docs.rs/bevy/0.14.2/bevy/prelude/struct.Text2dBundle.html) to display the value of each tile. `Text` uses "sections" to define updateable areas of content, so we use `[from_section](https://docs.rs/bevy/0.14/bevy/prelude/struct.Text.html#method.from_section)`. We initialize the displayed string to `"2"` because that will always be the starting value, although later on this value will update so quickly we won't even notice the initial value.

We also insert the `TileText` label component on the entity we got from spawning the `Text2dBundle` so we can find it later for all tiles.

```rust
.with_children(|child_builder| {
    child_builder.spawn((
        Text2dBundle {
            text: Text::from_section(
                "2",
                TextStyle {
                    font_size: 40.0,
                    color: Color::BLACK,
                    ..default()
                },
            )
            .with_justify(JustifyText::Center),
            transform: Transform::from_xyz(
                0.0, 0.0, 1.0,
            ),
            ..default()
        },
        TileText,
    ));
});
```

The `Text` is using the default font, which in 0.14 happens to be [Fira Mono](https://fonts.google.com/specimen/Fira+Mono).

![text on tiles](https://res.cloudinary.com/dilgcuzda/image/upload/v1727795683/workshops/2048-with-bevy-ecs/bevy-0.14/010-text-on-tiles.avif)