011 - Syncing Tile Text with Point Values
As we progress into building out the movement and merging functionality for tiles, we'll want to be able to see the point values as they change.

To do this we can write a system that is responsible for updating the tile text when the point values change.

Starting off, we'll change the text section we display by default to `4`.

```rust
Text::from_section("4",...
```

Running the game with `cargo run` will now show a board with tiles that display `4`.

![tiles rendering 4](https://res.cloudinary.com/dilgcuzda/image/upload/v1727796111/workshops/2048-with-bevy-ecs/bevy-0.14/011-tiles-displaying-4.avif)

The point of this is that we didn't change the underlying `Points` value, so if our system is successful the board will change from displaying 4, the rendered default, to 2, the underlying points value, allowing us to see that it is working.
r

The system is going to run once every frame, so we can use `.add_systems` to add the `sync_tile_points` system to the `Update` schedule.

```rust
.add_systems(Update, sync_tile_points)
```

The `sync_tile_points` system is going to query for all of the `Text` components that we labelled with `TileText`, alongside their `Transform` and their `Parent`.

Bevy inserts `Parent` and `Children` components for us. They store Entity ids which allows us to look for related entities from the top-down or from the bottom-up. We'll look up one level, also known as the `Parent`.

Bevy is a bit like a database, so we can use the `Parent` entity id as a key `.get` the right `Points` value from any given `Query`, such as this `Query<&Points>`.

```rust
fn sync_tile_points(
    mut texts: Query<
        (&mut Text, &mut Transform, &Parent),
        With<TileText>,
    >,
    tiles: Query<&Points>,
) {
    for (mut text, mut transform, parent) in &mut texts {
        let Ok(points) = tiles.get(parent.get()) else {
            warn!("An entity with TileText should have a parent with a Points component");
            continue;
        };

        text.sections[0].value = points.value.to_string();
        *transform = transform.with_scale(Vec3::splat(
            1.0 / points.value.to_string().len() as f32,
        ));
    }
}
```

We can use Iterators to iterate over all of the tiles, destructuring the text, transform, and parent from the tuple.

Then we'll use `tiles` as a database, and use the parent's entity id as the argument to `tiles.get()`. This returns a `Result` because it could fail.

In the event that it fails, we use a pattern that has become pretty popular in the Bevy ecosystem. We can destructure the `Result` using let-else syntax. If the `.get` succeeds and returns an `Ok` value, then we continue. If it doesn't, the `else` block will run.

In our case, we'll `warn` using Bevy's logging infrastructure before continuing to process the rest of the `texts`.

```rust
let Ok(points) = tiles.get(parent.get()) else {
    warn!("An entity with TileText should have a parent with a Points component");
    continue;
};
```

We can then immediately use the `points.value` to set the `value` of the first text section. `points.value` is a `u32`, so we need to turn that into a `String` to use it as the text section value.

```rust
text.sections[0].value = points.value.to_string();
```

Running the game now will show the underlying points value instead of the `4`s we have set.

![rendering 2s](https://res.cloudinary.com/dilgcuzda/image/upload/v1727927350/workshops/2048-with-bevy-ecs/bevy-0.14/011-rendering-2s.avif)

but there is one problem. The text will end up being up to four characters long (2048).

![overflowing text](https://res.cloudinary.com/dilgcuzda/image/upload/v1727927533/workshops/2048-with-bevy-ecs/bevy-0.14/011-overflowing-text.avif)

So finally, we include a neat bit of code to scale the size of the text to fit in the boxes we've built. The `Transform` component lets us add a scale using [`with_scale`](https://docs.rs/bevy/0.14/bevy/transform/components/struct.Transform.html#method.with_scale).

We'll set this value to `1` over the length of the text. So a 2 will be `1/1`, a 16 will be `1/2`, a 256 will be `1/3` and 2048 will be `1/4`.

```rust
*transform = transform.with_scale(Vec3::splat(
    1.0 / points.value.to_string().len() as f32,
));
```

The scaling we've chosen is arbitrary. You could also change the font size or pick different numbers if you want to.

![non overflowing](https://res.cloudinary.com/dilgcuzda/image/upload/v1727927549/workshops/2048-with-bevy-ecs/bevy-0.14/011-non-overflowing.avif)

Remember to change the `Points` default value to 2 and the `4` in our text string back to 2 as well.

With the system running, you'll see the tiles display 2 again which means everything is working.

![fitting text](https://res.cloudinary.com/dilgcuzda/image/upload/v1727928172/workshops/2048-with-bevy-ecs/bevy-0.14/011-fitting-text.avif)