012 - listening-for-keyboard-input
In 2048 the user presses arrow keys and all of the tiles shift in that direction, merging if possible. To represent this we'll write a new system to handle user input and translate it into tile movement. This new system will be called `board_shift`.

Keyboard input in Bevy is represented as a resource, so we'll ask Bevy for `Res<Input<KeyCode>>`. `[Input](https://docs.rs/bevy/0.10.0/bevy/input/prelude/struct.Input.html)` has a few functions for grabbing which keys are being pressed. `[input.pressed](https://docs.rs/bevy/0.10.0/bevy/input/prelude/struct.Input.html#method.pressed)` will match against key presses every game tick. This is useful for games where you have a character that is running and need to constantly apply movement.

We have a 2d board game, so we'll use `[just_pressed](https://docs.rs/bevy/0.10.0/bevy/input/prelude/struct.Input.html#method.just_pressed)` instead, which will only match once when a key is pressed. We can apply this function up to four times to figure out which key is pressed.

```rust
fn board_shift(input: Res<ButtonInput<KeyCode>>) {
    if input.just_pressed(KeyCode::ArrowLeft) {
        dbg!("left");
    } else if input.just_pressed(KeyCode::ArrowRight) {
        dbg!("right");
    } else if input.just_pressed(KeyCode::ArrowDown) {
        dbg!("down");
    } else if input.just_pressed(KeyCode::ArrowUp) {
        dbg!("up");
    }
}
```

Adding this system to our `App` will listen to pressed keys and log out which key was pressed. Since this system can run at the same time as `render_tile_points`, we can change our `add_systems` to use a tuple.

```rust
.add_systems(Update, (sync_tile_points, board_shift))
```

We'll translate the incoming key presses into something we can `match` on for the rest of our application. This is a useful concept because in larger games many different keys can trigger the same action. Consider WASD vs arrow keys, or gamepad D-pad button presses. We want one way to respond to all of these in our application without needing to care which was actually pressed.

We'll create a new `enum` call `BoardShift` to hold the various directions the board can shift in.

```rust
enum BoardShift {
    Left,
    Right,
    Up,
    Down,
}
```

To turn a `KeyCode` into a `BoardShift` we can implement the `TryFrom` trait which is in the Rust standard library. We need to bring the trait into scope to be able to implement it (or use it).

```rust
use std::convert::TryFrom;
```

Then we can implement `TryFrom` for a shared reference to a `KeyCode`. We implement the trait on a shared reference because we don't need to mutate it, only read it.

To complete the implementation of the trait we have to provide a type for the error if something goes wrong and a function called `try_from` to do the conversion. In this case we use a `&'static string` for the `Error` because we'll be writing a hardcoded string.

The `try_from` function accepts a shared reference to a `KeyCode`, which is the same type we're using in the impl type above. The return type is a `Result`, because this operation may fail so we want to be able to return `Ok()` or `Err()` depending on what happens. `Self` refers to the type we're implementing the trait for, `BoardShift` while `Self::Error` refers to the associated type `Error` that we defined. An alternate way of writing this type signature would be `Result<BoardShift, &'static str>`.

The [`KeyCode`](https://docs.rs/bevy/0.14/bevy/input/keyboard/enum.KeyCode.html) enum contains over 100 possible keys that can be pressed, while we only care about four of them. We can match on the ones we care about, while using `_` to ignore all of the others.

The conversion is one-to-one in that `KeyCode::ArrowLeft` becomes `BoardShift::Left`. We do have to wrap the successful values in `Ok` and the error value in `Err` to match the type signature.

```rust
impl TryFrom<&KeyCode> for BoardShift {
    type Error = &'static str;

    fn try_from(
        value: &KeyCode,
    ) -> Result<Self, Self::Error> {
        match value {
            KeyCode::ArrowLeft => Ok(BoardShift::Left),
            KeyCode::ArrowUp => Ok(BoardShift::Up),
            KeyCode::ArrowRight => Ok(BoardShift::Right),
            KeyCode::ArrowDown => Ok(BoardShift::Down),
            _ => Err("not a valid board_shift key"),
        }
    }
}
```

With this implementation, whenever we have a `KeyCode` we can use `BoardShift::try_from` to attempt to convert it to a board shift direction.

```rust
BoardShift::try_from(key_code)
```

In our `board_shift` system then, we can access an iterator (called [`get_just_pressed`](https://docs.rs/bevy/0.14/bevy/input/struct.ButtonInput.html#method.get_just_pressed)) over all of the keys that were just pressed. This makes it clear that multiple keys *can* be pressed, even though we're about to find the first match anyway.

[`.find_map`](https://doc.rust-lang.org/std/iter/trait.Iterator.html#method.find_map) is an interesting function. Because Rust has the `Option` type we can combine `.find` and `.map` into a single function, wrapping the values we want in `Some` and replacing the values we don't with `None`. More directly, we are going to find the first `KeyCode` that can be converted into a `BoardShift` and then change that value into the `BoardShift` value, skipping all others.

```rust
let shift_direction =
    input.get_just_pressed().find_map(
        |key_code| BoardShift::try_from(key_code).ok(),
    );
```

`.find_map` accepts a closure that gives us a single item from the iterator at a time. `BoardShift::try_from` allows us to try to convert that item into a `BoardShift` position. `try_from` returns a `Result` and `find_map` wants an `Option`, so we can discard the error (which we don't care about anyway) and convert the `Result` into an `Option` with `.ok()`.

Once we've run `find_map` we can use let-else syntax to destructure the value, or early-return from the system if the user hasn't pressed a valid key.

```rust
fn board_shift(input: Res<ButtonInput<KeyCode>>) {
    let Some(shift_direction) =
        input.get_just_pressed().find_map(|key_code| {
            BoardShift::try_from(key_code).ok()
        })
    else {
        return;
    };

    dbg!(shift_direction);
}
```

Running this code with `cargo run` shows us which shift direction we've pressed, if we've pressed one.

```rust
[src/main.rs:221:5] shift_direction = Up
[src/main.rs:221:5] shift_direction = Left
[src/main.rs:221:5] shift_direction = Down
[src/main.rs:221:5] shift_direction = Right
[src/main.rs:221:5] shift_direction = Up
[src/main.rs:221:5] shift_direction = Left
```