015 - updating tile positions with transforms
When the tile `Position` components are updated, the physical position on the board is not, because we don't have a system to take care of translating `Position` values into physical positions on the screen.

We'll write a `render_tiles` system and add it to the always-running systems.

```rust
.add_systems(Update, (
    sync_tile_points,
	board_shift,
	sync_tile_positions
))
```

The `render_tiles` system will mutate the `[Transform](https://docs.rs/bevy/0.14/bevy/prelude/struct.Transform.html)` component on the tile, and read from the `Position` component. The biggest new piece of this query is `[Changed](https://docs.rs/bevy/0.14/bevy/ecs/prelude/struct.Changed.html)<Position>` which will filter the tiles in our query down to only the ones whose `Position` components have changed. Using this query means we don't have to do any complicated state maintenance across game ticks to make sure we're only updating when something changes.

We'll also query the `Board` to take advantage of the `board.grid_to_world_position` function we wrote.

```rust
fn sync_tile_positions(
    mut tiles: Query<
        (&mut Transform, &Position),
        Changed<Position>,
    >,
    board: Res<Board>,
) {
    for (mut transform, pos) in tiles.iter_mut() {
        transform.translation.x =
            board.grid_to_world_position(pos.0.x);
        transform.translation.y =
            board.grid_to_world_position(pos.0.y);
    }
}
```

Iterating over the tiles with `&mut tiles` allows us to check to see if the `Position` changed, and if it did then apply the physical location change to the translation.

`Transform`s are a very common Bevy component made up of a translation, rotation, and scale. Any displayable entity in our app will have a `Transform` component. In our case we only care about the translation, so we'll update the x and y values here.

Now when pressing left, all of the tiles will move to the left and potentially merge if applicable.

![starting board state](https://res.cloudinary.com/dilgcuzda/image/upload/v1728023566/workshops/2048-with-bevy-ecs/bevy-0.14/015-starting-state.avif)

![tiles shifted left](https://res.cloudinary.com/dilgcuzda/image/upload/v1728023578/workshops/2048-with-bevy-ecs/bevy-0.14/015-shifted-left.avif)